requirements.txt
PyQt5>=5.15.9  # 图形界面库
opencv-python>=4.8.1.78  # 图像处理库
Pillow>=10.1.0  # 图像操作库
numpy>=1.26.2  # 数值计算库    